Separate main and app files
